Initiated the refactoring task for this project. New Gmail UI Login completed.
Page Object design for navigating to the Inbox.
