+++ autogmailpy/__init__.py
@@ -1,4 +1,3 @@
+from autogmailpy.pages.gmail_inbox_page import *
+from autogmailpy.pages.gmail_login_page import *
 __author__ = 'Jorge'
-
-from autogmailpy.pages.gmail_inbox import *
-from autogmailpy.pages.gmail_login import *
